Potential fix for code scanning alert no. 2: Implicit narrowing conversion in compound assignment [skip ci]

+++ spigotv-api/src/main/java/co/aikar/timings/TimingData.java
@@ -48,7 +48,7 @@ class TimingData {
     long lagTotalTime = 0;
 
     int curTickCount = 0;
-    int curTickTotal = 0;
+    long curTickTotal = 0;
 
     TimingData(int id) {
         this.id = id;
